% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{flatten_json}
\alias{flatten_json}
\title{Flatten JSON}
\usage{
flatten_json(x)
}
\arguments{
\item{x}{a json-like object}
}
\value{
a named vector
}
\description{
This function flattens a json like objects
into named vectors
}
