% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_time_log}
\alias{get_time_log}
\title{Get Time Log}
\usage{
get_time_log(df, k, type = "start")
}
\arguments{
\item{df}{dataframe for processing}

\item{k}{the kth window}

\item{type}{log type (start of end)}
}
\value{
time bound for logging
}
\description{
Get Time Log
}
